dl-fldigi waterfall loading
Since dl-fldigi writes to the same WF image all
the time, sometimes browser gets the image file
when it is incomplete and can't be displayed.
This commit tries to address this issue by 
checking the file on the server-side, then sending
a base64 encoded image with Ajax request, if the
file is OK.
Probably too complicated, but seems to work.